[feat] 일급 객체 `Name` 분리
